backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
alembic==1.13.1
asyncpg==0.29.0
psycopg2-binary==2.9.9
pydantic==2.5.3
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
google-genai>=1.0.0
google-auth-oauthlib>=1.2.0
httpx==0.26.0
python-dotenv==1.0.0
pgvector==0.2.4
numpy>=1.26.0
slowapi==0.1.9
motor>=3.3.0
sentence-transformers>=2.3.0
